Create a module for candidates
