+++ beets/autotag/__init__.py
@@ -7,18 +7,15 @@ from importlib import import_module
 # Parts of external interface.
 from beets.util.deprecation import deprecate_for_maintainers, deprecate_imports
 
-from .distance import Distance, distance, string_dist, track_distance
-from .hooks import AlbumInfo, Info, TrackInfo, correct_list_fields
-from .match import (
+from .candidates import (
     AlbumCandidates,
-    AlbumMatch,
     Candidates,
-    Match,
     Recommendation,
     TrackCandidates,
-    TrackMatch,
-    assign_items,
 )
+from .distance import Distance, distance, string_dist, track_distance
+from .hooks import AlbumInfo, Info, TrackInfo, correct_list_fields
+from .match import AlbumMatch, Match, TrackMatch, assign_items
 from .source import Source
